Реализация м1 - микросервиса для работы с JWT и Spring Security

+++ src/main/resources/db/migration/V3__create_confirm_token_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS confirm_tokens
 (
     id            SERIAL          PRIMARY KEY NOT NULL,
     token         uuid            NOT NULL UNIQUE,
-    person_id     BIGINT          NOT NULL,
+    person_id     int             NOT NULL,
     expired       TIMESTAMP       NOT NULL,
     CONSTRAINT person_id_fk FOREIGN KEY (person_id) REFERENCES person (id)
 );
